Economy Rush Update
Economy Rush bug fix with henchmen max at 20 for entire game.

Updated Henchman equations for EconomyRush as well as did some general cleanup. Adjusted distance requirements for Economy Rush so that AI could perform it on every location on Grove.

AI can now research Electrical Grid before Yoke if it is comfortable and currently in the process of researching L3.